git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4956@1512909 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/test/org/apache/solr/core/QueryResultKeyTest.java
@@ -38,7 +38,7 @@ public class QueryResultKeyTest extends SolrTestCaseJ4 {
     // the hashcode should be the same even when the list
     // of filters is in a different order
     
-    Sort sort = new Sort(new SortField("test", SortField.Type.BYTE));
+    Sort sort = new Sort(new SortField("test", SortField.Type.INT));
     List<Query> filters = new ArrayList<Query>();
     filters.add(new TermQuery(new Term("test", "field")));
     filters.add(new TermQuery(new Term("test2", "field2")));
